% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getMethylSet}
\alias{getMethylSet}
\title{Export the simulateMethObject to a MethylSet to be used in the
R/Bioconductor package \code{minfi}.}
\usage{
getMethylSet(simulateMethObject)
}
\arguments{
\item{simulateMethObject}{Must be an object created from \code{simulateMeth}}
}
\value{
A \code{MethylSet} object
}
\description{
This function takes in the simulateMethObject and
returns a MethylSet object which can be used in the minfi package.
Can use functions from mini such as \code{getMeth}, \code{getUnmeth},
\code{getBeta}, \code{getM}, \code{getCN}. Because it is simulated data,
there is no manifest and no preprocessing in the MethylSet returned.
}
\examples{
methTruth <- simulateMethTruth(nProbes = 2e4, nGroups = 2)
sim <- simulateMeth(methTruth, meth.platform = "methArrays",
                   nSamps = 5, nMol = 10^6)
mset <- getMethylSet(sim)
}
\author{
Stephanie Hicks
}

