% Generated by roxygen2 (4.0.2): do not edit by hand
\name{langmuirGExArrays}
\alias{langmuirGExArrays}
\title{Langmuir parameters for gene expression using arrays}
\usage{
langmuirGExArrays(nProbes, nSamps, nGroups, mua = NULL, siga = NULL,
  mub = NULL, sigb = NULL, muOpt = NULL, sigOpt = NULL, muBG = NULL,
  sigBG = NULL, muERR = NULL, sigERR = NULL)
}
\arguments{
\item{nProbes}{number of probes}

\item{nSamps}{number of samples in each group}

\item{nGroups}{number of groups}

\item{mua}{hyperparameter: mean of log normal distribution for a}

\item{siga}{hyperparameter: variance of log normal distribution for a}

\item{mub}{hyperparameter: mean of log normal distribution for b}

\item{sigb}{hyperparameter: variance of log normal distribution for b}

\item{muOpt}{hyperparameter: mean of log normal distribution for optical
noise}

\item{sigOpt}{hyperparameter: variance of log normal distribution for
optical noise}

\item{muBG}{hyperparameter: mean of log normal distribution for d}

\item{sigBG}{hyperparameter: variance of log normal distribution for d}

\item{muERR}{hyperparameter: mean of log normal distribution for
measurement error}

\item{sigERR}{hyperparameter: variance of log normal distribution for
measurement error}
}
\value{
A list with elements
\item{gID}{Group number ID}
\item{a}{Simulated a values for the Langmuir Adsorption model}
\item{b}{Simulated b values for the Langmuir Adsorption model}
\item{d}{Simulated d values for the Langmuir Adsorption model}
\item{opt}{Simulated optical noise}
\item{errPM}{Simulated measurement error }
}
\description{
These are the parameters used to simulate gene expression
arrays. The parmeters are used in the Langmuir adsorption model.
}
\examples{
langmuirGExArrays(nProbes = 50000, nSamps = 4, nGroups = 2)
}
\author{
Stephanie Hicks
}

