% Generated by roxygen2 (4.0.2): do not edit by hand
\name{langmuirTrans}
\alias{langmuirTrans}
\title{Langmuir adsorption transformation}
\usage{
langmuirTrans(x, a, b, d)
}
\arguments{
\item{x}{expected number of methylated molecules after PCR and
bisulfite-sequencing}

\item{a}{intensity from scanner}

\item{b}{scale parameter}

\item{d}{background noise}
}
\description{
To model DNA methylation and gene expression from arrays
this function represents the saturation reached in the arrays.
}
\author{
Stephanie Hicks
}

