% Generated by roxygen2 (4.0.2): do not edit by hand
\name{list.GEx.platforms}
\alias{list.GEx.platforms}
\title{List of platforms available to simulate gene expression data}
\usage{
list.GEx.platforms()
}
\value{
List of platforms available to simulate gene expression data
}
\description{
List of platforms available to simulate gene expression data
}
\examples{
list.GEx.platforms()
}
\author{
Stephanie Hicks
}

