% Generated by roxygen2 (4.0.2): do not edit by hand
\name{list.meth.platforms}
\alias{list.meth.platforms}
\title{List of platforms available to simulate DNA methylation data}
\usage{
list.meth.platforms()
}
\value{
List of platforms available to simulate gene expression data
}
\description{
List of platforms available to simulate DNA methylation data
}
\examples{
list.meth.platforms()
}
\author{
Stephanie Hicks
}

