% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pickLangmuir}
\alias{pickLangmuir}
\title{Simulate parameters from the Langmuir Adsorption model}
\usage{
pickLangmuir(objectType, typePlatform, nProbes, nSamps, nGroups, mua = NULL,
  siga = NULL, mub = NULL, sigb = NULL, muOpt = NULL, sigOpt = NULL,
  muBG = NULL, sigBG = NULL, muERR = NULL, sigERR = NULL)
}
\arguments{
\item{objectType}{Must be "simulateGExTruthObject" or
"simulateMethTruthObject"}

\item{typePlatform}{Must specify a platform from \code{list.GEx.platforms}
or \code{list.meth.platforms}.}

\item{nProbes}{number of probes}

\item{nSamps}{number of samples in each group}

\item{nGroups}{number of groups}

\item{mua}{hyperparameter: mean of log normal distribution for a}

\item{siga}{hyperparameter: variance of log normal distribution for a}

\item{mub}{b hyperparameter: mean of log normal distribution for b}

\item{sigb}{b hyperparameter: variance of log normal distribution for b}

\item{muOpt}{hyperparameter: mean of log normal distribution for
optical noise}

\item{sigOpt}{hyperparameter: variance of log normal distribution for
optical noise}

\item{muBG}{hyperparameter: mean of log normal distribution for d}

\item{sigBG}{hyperparameter: variance of log normal distribution for d}

\item{muERR}{hyperparameter: mean of log normal distribution for
measurement error}

\item{sigERR}{hyperparameter: variance of log normal distribution for
measurement error}
}
\value{
The output of this function is the output directly reported from
the functions \code{langmuirGExArrays} and \code{langmuirMethArrays}.
}
\description{
This is a helper function that points to the
functions which contain the default parameters for each platform
available in this package. This function can simulate
parameters from the Langmuir model for each of the platforms available
\code{list.GEx.platforms} and \code{list.meth.platforms}.
}
\examples{
ps <- pickLangmuir(objectType = "simulateMethTruthObject",
                   typePlatform = "methArrays", nProbes = 1e5,
                   nSamps = 5, nGroups = 2)
}
\author{
Stephanie Hicks
}

