% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pickLogNormal}
\alias{pickLogNormal}
\title{Simulate the log normal values}
\usage{
pickLogNormal(N, muHyp = NULL, sigHyp = NULL)
}
\arguments{
\item{N}{number of log normal means to simulate}

\item{muHyp}{If NULL (default), hyperparameters for mean of log normal
are simulated using Uniform(0, 6). Otherwise, user must supply muHyp.}

\item{sigHyp}{If NULL (default), hyperparameters for variance of log
normal are simulated using Uniform(0.1, 2). Otherwise, user must
supply sigHyp.}
}
\value{
The output is a vector of log normal values.
}
\description{
Helper function to simulate the log normal mean values with
varying hyperparameters (means and variances). To be used as input for
the means of the Poisson distribution.
}
\examples{
pickLogNormal(N = 1000)
}
\author{
Stephanie Hicks
}

