% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pickNormalMix}
\alias{pickNormalMix}
\title{Simulate a mixture of normal distributions}
\usage{
pickNormalMix(N, nMix = NULL, propMix = NULL, muMix = NULL,
  sigMix = NULL)
}
\arguments{
\item{N}{number of observations to simulate from the mixture of normals}

\item{nMix}{number of mixtures. Default is 3.}

\item{propMix}{mixture proportions. If NULL (default), a mixture
distribution with propMix = c(0.4, 0.2, 0.4) is simulated. Otherwise
user must supply propMix. Length should equal nMix.}

\item{muMix}{mean of nMix distributions. If NULL (default), a mixture
distribution with muMix = c(-3, 1, 3) is simulated. Otherwise user must
supply propMix. Length should equal nMix.}

\item{sigMix}{variance of nMix distributions. If NULL (default), a mixture
distribution with sigMix = c(3, .4, 3) is simulated. Otherwise user must
supply sigMix. Length should equal nMix.}
}
\value{
A list with elements
\item{trueParams}{A data frame with the input parameters used to simulate
the mixture distribution (propMix, muMix and sigMix)}
\item{simObs}{A data frame containing trueMix (indicator specifying which
distribution the observation came from) and normObs (the observed value
from the mixture distribution).}
}
\description{
Helper function to simulate the normal mixtures with varying
hyperparameters (means and variances).
}
\examples{
pickNormalMix(N = 10)
}
\author{
Stephanie Hicks
}

