% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plotGEx}
\alias{plotGEx}
\title{Plot the simulated gene expression arrays or RNA-Sequencing}
\usage{
plotGEx(simulateGExObject)
}
\arguments{
\item{simulateGExObject}{Must be an object created from \code{simulateGEx}}
}
\description{
Plot the simulated gene expression arrays or RNA-Sequencing.
}
\examples{
geneTruth <- simulateGExTruth(nGenes = 2e4, nGroups = 2,
                             pDiff = 0.05, foldDiff = 5)
sim <- simulateGEx(geneTruth, GEx.platform = "GExArrays", nSamps = 5)
plotGEx(sim)
}
\author{
Stephanie Hicks
}

