% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plotGExTruth}
\alias{plotGExTruth}
\title{Plot the \code{simulateGExTruthObject} from \code{simulateGExTruth}}
\usage{
plotGExTruth(simulateGExTruthObject)
}
\arguments{
\item{simulateGExTruthObject}{Must be an object created from
\code{simulateGExTruth}.}
}
\value{
The distribution of simulated RNA molecules for each gene in
each group.
}
\description{
This function plots the simulated RNA molecules on the
log 2 scale from the \code{simulateGExTruthObject}.
}
\examples{
geneTruth <- simulateGExTruth(nGenes = 2e4, nGroups = 2,
                              pDiff = 0.05, foldDiff = 5)
plotGExTruth(geneTruth)
}

