% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plotMeth}
\alias{plotMeth}
\title{Plot the simulated DNA methylation arrays}
\usage{
plotMeth(simulateMethObject)
}
\arguments{
\item{simulateMethObject}{Must be an object created from \code{simulateMeth}}
}
\description{
This function requires the simulateMethObject from
\code{simulateMeth}. Boxplots and density plots are plotted showing the
distribution of DNA methylation.
}
\examples{
methTruth <- simulateMethTruth(nProbes = 2e4, nGroups = 2,
                               pDiff = 0.05, pUp = 0.80)
sim <- simulateMeth(methTruth, meth.platform = "methArrays",
                     nSamps = 5, nMol = 10^6)
plotMeth(sim)
}
\author{
Stephanie Hicks
}

