% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plotMethTruth}
\alias{plotMethTruth}
\title{Plot the \code{simulateMethTruthObject} from \code{simulateMethTruth}}
\usage{
plotMethTruth(simulateMethTruthObject)
}
\arguments{
\item{simulateMethTruthObject}{Must be an object created from
\code{simulateMethTruth}.}
}
\value{
Two plots: One is a histogram of mixture model of three normal
distributions. The second is the density of the transformed methylation data
scaled between 0 and 1.
}
\description{
This function plots the simulate DNA meylation data from the
\code{simulateMethTruthObject}.
}
\examples{
methTruth <- simulateMethTruth(nProbes = 2e4, nGroups = 2,
                               pDiff = 0.05, pUp = 0.80)
plotMethTruth(methTruth)
}

