% Generated by roxygen2 (4.0.2): do not edit by hand
\name{simulateGEx}
\alias{simulateGEx}
\title{Simulate observed gene expression data for a set of samples}
\usage{
simulateGEx(simulateGExTruthObject, GEx.platform = "GExArrays", nSamps,
  usePCR = FALSE, nPCRcycles = 15, verbose = TRUE, mua = NULL,
  siga = NULL, mub = NULL, sigb = NULL, muOpt = NULL, sigOpt = NULL,
  muBG = NULL, sigBG = NULL, muERR = NULL, sigERR = NULL)
}
\arguments{
\item{simulateGExTruthObject}{Must be an object created from
\code{simulateGExTruth}.}

\item{GEx.platform}{Must specify a platform from \code{list.GEx.platforms}.}

\item{nSamps}{number of samples in each group}

\item{usePCR}{TRUE/FALSE option. Default is FALSE.
If TRUE, amplify RNA fragments with PCR.}

\item{nPCRcycles}{number of PCR cycles. Default is 15.}

\item{verbose}{TRUE/FALSE option to print details about simulation.
Default is TRUE.}

\item{mua}{hyperparameter: mean of log normal distribution for a}

\item{siga}{hyperparameter: variance of log normal distribution for a}

\item{mub}{b hyperparameter: mean of log normal distribution for b}

\item{sigb}{b hyperparameter: variance of log normal distribution for b}

\item{muOpt}{hyperparameter: mean of log normal distribution for
optical noise}

\item{sigOpt}{hyperparameter: variance of log normal distribution for
optical noise}

\item{muBG}{hyperparameter: mean of log normal distribution for d}

\item{sigBG}{hyperparameter: variance of log normal distribution for d}

\item{muERR}{hyperparameter: mean of log normal distribution for
measurement error}

\item{sigERR}{hyperparameter: variance of log normal distribution for
measurement error}
}
\value{
A list of elements
\item{objectType}{a string specifying the type of object.}
\item{typePlatform}{platform used to simulate the gene expression.}
\item{usePCR}{A TRUE/FALSE object specifying if PCR was used to amplify
the RNA counts.}
\item{nGenes}{number of genes}
\item{nProbes}{number of probes}
\item{nSamps}{number of samples}
\item{nGroups}{number of groups}
\item{params}{list of elements reported from \code{pickLangmuirParameters}}
\item{geneNames}{gene names (or probeset names)}
\item{nProbesPerGene}{number of probes per gene (or probeset)}
\item{pd}{phenoData containing information about the simulated samples}
\item{PM}{the simulated gene expression data}
}
\description{
This function simulates observed gene expression data
based on the set of platforms listed in \code{list.GEx.platforms}.
The gene expression arrays are simulated using the Langmuir adsorption
model.
}
\examples{
geneTruth <- simulateGExTruth(nGenes = 2e4, nGroups = 2,
                             pDiff = 0.05, foldDiff = 5)
sim <- simulateGEx(geneTruth, GEx.platform = "GExArrays", nSamps = 5)
}
\author{
Stephanie Hicks
}

