% Generated by roxygen2 (4.0.2): do not edit by hand
\name{simulateGExTruth}
\alias{simulateGExTruth}
\title{Simulate true gene expression data for a set of groups}
\usage{
simulateGExTruth(nGenes, nGroups = 2, muHyp = NULL, sigHyp = NULL,
  zeroWeight = 0.1, pDiff = 0.05, foldDiff = 5, scaleCounts = 1,
  verbose = TRUE)
}
\arguments{
\item{nGenes}{number of genes to simulate.}

\item{nGroups}{number of groups. Default is 2.}

\item{muHyp}{hyperparameters for mean of log normal. See
\code{pickLogNormal} for default values.}

\item{sigHyp}{hyperparameters for variance of log normal. See
\code{pickLogNormal} for default values.}

\item{zeroWeight}{proportion of zero-inflated counts. Default is 0.10.}

\item{pDiff}{percent of probes different relative to Group 1. If
\code{nGroups == 1}, pDiff should be 0. If \code{nGroups > 1}, the
length of pDiff should be equal to \code{nGroups - 1}. Default
is \code{pDiff = 0.05}.}

\item{foldDiff}{fold difference relative to Group 1. If
\code{nGroups == 1}, foldDiff is ignored. If \code{nGroups > 1}, the
length of foldDiff should be equal to \code{nGroups - 1}. Default is
\code{foldDiff = 5}.}

\item{scaleCounts}{To plot gene expression on log2 scale, pick a
small amount to scale the counts by. Default is 1.}

\item{verbose}{TRUE/FALSE argument specifying if verbose messages
should be returned or not. Default is TRUE.}
}
\value{
A list with elements
\item{objectType}{A string specifying the type of object.}
\item{nGenes}{Number of genes simulated.}
\item{zeroWeight}{Proportion of zero-inflated counts used in simulation.}
\item{pDiff}{Percent of probes different relative to Group 1.}
\item{foldDiff}{fold difference relative to Group 1.}
\item{scaleCounts}{An positive real valued number specifying an additive
scale for the RNA counts to be able to plot on the log2 scale.}
\item{genesDiffInd}{Boolean array referencing which probes are different
relative to Group 1.}
\item{geneRange}{Data frame containing the simulated RNA molecules
for each gene and for each group.}
}
\description{
This function simulates true gene expression data.
The gene expression arrays are simulated using a zero inflated
log-normal poisson distribution to represent the number or RNA
molecules in a sample.
}
\examples{
geneTruth <- simulateGExTruth(nGenes = 2e4, nGroups = 2,
                               pDiff = 0.05, foldDiff = 5)
}
\author{
Stephanie Hicks
}
\seealso{
\code{\link{simulateMethTruth}}
}

