% Generated by roxygen2 (4.0.2): do not edit by hand
\name{simulateMeth}
\alias{simulateMeth}
\title{Simulate observed DNA methylation data for a set of samples}
\usage{
simulateMeth(simulateMethTruthObject, meth.platform, nSamps, nMol = NULL,
  verbose = TRUE, mua = NULL, siga = NULL, mub = NULL, sigb = NULL,
  muOpt = NULL, sigOpt = NULL, muBG = NULL, sigBG = NULL,
  muERR = NULL, sigERR = NULL)
}
\arguments{
\item{simulateMethTruthObject}{Must be an object created from
\code{simulateMethTruth}.}

\item{meth.platform}{Must specify a platform from
\code{list.meth.platforms}.}

\item{nSamps}{number of samples in each group}

\item{nMol}{number of molecules after amplification}

\item{verbose}{TRUE/FALSE option to print details about simulation.
Default is TRUE.}

\item{mua}{hyperparameter: mean of log normal distribution for a}

\item{siga}{hyperparameter: variance of log normal distribution for a}

\item{mub}{b hyperparameter: mean of log normal distribution for b}

\item{sigb}{b hyperparameter: variance of log normal distribution for b}

\item{muOpt}{hyperparameter: mean of log normal distribution for
optical noise}

\item{sigOpt}{hyperparameter: variance of log normal distribution for
optical noise}

\item{muBG}{hyperparameter: mean of log normal distribution for d}

\item{sigBG}{hyperparameter: variance of log normal distribution for d}

\item{muERR}{hyperparameter: mean of log normal distribution for
measurement error}

\item{sigERR}{hyperparameter: variance of log normal distribution for
measurement error}
}
\value{
A list of elements
\item{objectType}{a string specifying the type of object.}
\item{typePlatform}{platform used to simulate the gene expression.}
\item{nProbes}{number of probes}
\item{nSamps}{number of samples}
\item{nGroups}{number of groups}
\item{params}{list of elements reported from \code{pickLangmuirParameters}}
\item{pd}{phenoData containing information about the simulated samples}
\item{meth}{the simulated raw methylated data}
\item{unmeth}{the simulated raw unmethylated data}
}
\description{
This function simulates observed DNA methylation data
based on the set of platforms listed in \code{list.meth.platforms}.
The DNA methylation is simulated using the Langmuir adsorption model.
}
\examples{
methTruth <- simulateMethTruth(nProbes = 2e4, nGroups = 2,
                               pDiff = 0.05, pUp = 0.80)
sim <- simulateMeth(methTruth, meth.platform = "methArrays",
                    nSamps = 5, nMol = 10^6)
}
\author{
Stephanie Hicks
}

