% Generated by roxygen2 (4.0.2): do not edit by hand
\name{simulateMethTruth}
\alias{simulateMethTruth}
\title{Simulate true DNA methylation data for a set of groups}
\usage{
simulateMethTruth(nProbes, nGroups = 2, nMixtures = 3,
  propMixtures = NULL, muMixtures = NULL, sigMixtures = NULL,
  pDiff = 0.05, pUp = 0.8, verbose = TRUE)
}
\arguments{
\item{nProbes}{number of probes}

\item{nGroups}{number of groups. Default is 2.}

\item{nMixtures}{number of normal distributions to simulate. Default is 3.
Can be changed, but user must then supply mixing proportions (propMixtures),
mean (muMixtures) and variance (sigMixtures).}

\item{propMixtures}{weights or mixture proportions for the normal
distributions. Should be a vector of length nMixtures. See
\code{pickNormalMix} for default values.}

\item{muMixtures}{mean of the normal distributions. Should be a vector of
length nMixtures. See \code{pickNormalMix} for default values.}

\item{sigMixtures}{variance of the three normal distributions. Should be
a vector of length nMixtures.  See \code{pickNormalMix} for default values.}

\item{pDiff}{percent of probes different relative to Group 1. If
\code{nGroups == 1}, pDiff should be 0. If \code{nGroups > 1}, the
length of pDiff should be equal to \code{nGroups - 1}. Default
is \code{pDiff = 0.05}.}

\item{pUp}{proportion of pDiff probes that are methylated relative to
Group 1. If \code{nGroups == 1}, pUp is ignored. If \code{nGroups > 1}, the
length of pUp should be equal to \code{nGroups - 1}. Default is
\code{pUp = 0.80}.}

\item{verbose}{TRUE/FALSE argument specifying if verbose messages
should be returned or not. Default is TRUE.}
}
\value{
A list with elements
\item{objectType}{A string specifying the type of object.}
\item{nProbes}{Number of probes.}
\item{nGroups}{Number of groups.}
\item{trueParams}{Add me.}
\item{simObs}{Add me.}
\item{methDiffInd}{Add me.}
\item{methRange}{Add me.}
}
\description{
This function simulates true DNA methylation data. The true
DNA methylation data is simulated using a mixture of three normal
distributions to represent a probe that is not methylated, semi-methylated
and methylated. To scale the values between 0 and 1, the values are
transformed using an inverse-logit transformation. These scaled values
represent the true proportion of methylation at each probe.
}
\examples{
methTruth <- simulateMethTruth(nProbes = 2e4, nGroups = 2,
                                 pDiff = 0.05, pUp = 0.80)
}
\author{
Stephanie Hicks
}
\seealso{
\code{\link{simulateGExTruth}}
}

